Auto-completion demonstration
We can further improve our application to offer an auto-completion of
available Spring Boot versions. To do that we update first our template
to invoke the proper function

```
typeahead="version for version in bootVersions | filter:$viewValue | limitTo:8
```

Instead of the hard-coded version we can create an extra endpoint that
provides the full list of versions which uses our existing
SpringBootVersionService. We can now call our service by filling the
`springBootServiceUrl` and calling our endpoint in the fetchBootVersions
function.

We can also improve our app to cache this result and we can showcase how
we can build a custom health indicator for the remote URLs we are using.