Actuator demonstration
Once the actuator is in place, it is very easy to augment some of the
endpoints to expose application-specific meta-data.

The `info` endpoint exposes any arbitrary information about a running
application. In this commit we expose the name and version (taken from
the build information ) as well as git meta-data.

Metrics can be easily recorded (both individual values and counters). In
this commit, we record the number of times a particular version was
requested. We also use the new `@ExportMetricWriter` annotation to
export our metrics to a separate system. In this case, we simply expose
them to JMX but plenty of other options are available (redis, statsd,
etc).

Finally, we could also update the `health` endpoint to expose a custom
health indicator. For now, we're going to expose a simple one for
demonstration purposes.