Typesafe custom configuration demonstration
Our SpringBootVersionService only contacts the release repository in an
hard-coded fashion. It would be nice if we could customize the list of
repositories.

Spring Boot offers very nice auto-completion in the IDE and this feature
can be easily integrated in your own project. In this commit, we create
a dedicated configuration POJO for the Spring Boot version service and
use a map to index repository URLs by name. We enable the annotation
processor responsible to generate the meta-data at build time so that
the IDE can pick up our new key.